refactor: relocate target/diff_filter out of engine package
